增加文章阅读数，完善文章评论功能

+++ src/main/java/wang/kingweb/community/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
     }
 
     @PostMapping("/publish")
-    public String doPublish(@RequestParam(name = "id") Integer id,
+    public String doPublish(@RequestParam(name = "id") Long id,
                             @RequestParam(name = "title") String title,
                             @RequestParam(name = "description") String description,
                             @RequestParam(name = "tag") String tag,
